pdfsemanticsearchtest/requirements.txt
numpy==1.24.4
torch==2.2.2
sentence-transformers==2.6.1
huggingface_hub==0.23.0
faiss-cpu==1.7.4
pymupdf
